sys/backend/requirements.txt
# Core Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
orjson>=3.9.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
redis>=5.0.1

# Security
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.4

# Machine Learning
scikit-learn>=1.3.2
transformers>=4.35.2
torch>=2.0.0
shap>=0.44.1
lime>=0.2.0.1
sentence-transformers>=2.2.0
joblib>=1.3.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Async & Task Queue
celery[redis]>=5.3.4
asgiref>=3.7.2

# Monitoring & Logging
prometheus-client>=0.17.1
sentry-sdk[fastapi]>=1.30.0
structlog>=23.1.0

# API Documentation
description-content-type>=0.1.0
fastapi-pagination>=0.12.8

# Utilities
python-slugify>=8.0.1
pytz>=2023.3
tzlocal>=5.0.1

# File Storage
cloudinary>=1.36.0
python-magic>=0.4.27

# QR Codes
qrcode[pil]>=7.4.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.24.1
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Development
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
flake8>=6.1.0
pre-commit>=3.3.3